user testing ,, changed Cart name and tested it

+++ src/main/java/gov/iti/jets/genericDao/GenericDAO.java
@@ -7,11 +7,12 @@ public interface GenericDAO<T> {
   
     Set<T> findAll();
 
-    T findById(java.lang.Integer id);
+    T findById(Long id);
 
-    boolean delete(Integer id);
+    boolean delete(Long id);
 
     boolean delete(T entity);
 
     T update (T entity);
+
 }
